Simplified build process.
Migrations are now run automatically when building containers. As a result, the Diesel cli doesn't need to be installed on the host machine anymore. However, this does requre that the sql server is running to perform the migrations at build time.

For future builds, as long as the migrations folder is not modified, the migration step will be cached in the build process and so the build command can be executed without the sql container being up.