Mass deleting related posts is now much faster and should be immune to deadlocks.
Deleting related posts quickly can cause deadlocks due to cascade deletion of post relations. My previous strategy of dealing with this was to retry the deletion transaction a few times until one avoided a deadlock. However, this change implements a much better strategy, which is to lock a mutex while deleting posts with relations.