Fixed crashes after creating comments, pools, tags, and posts.
Oops! Making all those triggers deferred has some unintended side-effects. Namely, the statistics tables are only generated at the end of the transaction, so when the server tries to read a resource that requires a statistics tables it runs into trouble because it doesn't exist yet. The solution here is to split these transactions into a write part and a read part. That way the statistics tables are generated by the time they need to be read.

I've also added an admin command to reset database statistics if needed for migration or if they get out of sync due to implementation issues.

Changed most instances of std::Path::exists() to std::Path::try_exists().
Removed glob import of diesel::dsl.