Re-tuned image signature parameters.
This was motivated by me finding a case where the reverse image search did remarkably poorly for two identical images of different resolutions. Upon investigation, I found that the image signature process was placing the grid squares in different locations because it calculated the crop region to be different. I think this was because of some fine lines at the edges of the image which caused different spikes in the row and column deltas for different resolutions.

My conclusion is that the CROP_PERCENTILE parameter is overtuned. I've brought it down from 5 to 1, which is still enough to remove featureless borders. My next step is to think of a good way to introduce this image signature change to users, as it will require recomputing all image signatures. It's probably about time to look into speeding that up.