Improved speed of reverse search candidate filtering.
The signature::distance function was made considerably faster by:

1) Removing allocations from the uncompress function
2) Using a better algorithm for uncompress that doesn't rely on an intermediate array
3) Caching the decompression and norm computation of the content signature instead of recomputing them for every candidate signature

These improvements make the process of comparing the content signature to the candidate signature 5-10x faster.
This translates to an up to 10% improvement for reverse search for the first 5 uploads in a batch and an up to 1.5-2x improvement after the first 5 uploads in a batch.