Reverse search is now significantly faster.
I finally fixed the problem of reverse searches being slow for the first five images in a batch. Now, every image performs well right away. This speed up is significant: as much as 10x for small images (at least on my machine).

It turned out the problem was that Postgres was generated bad query plans for the first few reverse searches. Instead of performing an index scan, it would instead opt for a slow sequential scan for some reason. By disabling sequential scans for the candidate signature search, I was able to force Postgres to always use an index scan.