Conversion script now converts snapshots and post notes!
Finally figured out how to deserialize python objects using a SQL script, but it involves added an extra step to the conversion process.

We need to use python to deserialize the objects, but enabling python in PostgreSQL is not trivial. The extension needs some extra packages to work, so I've added the install_plptyhon3u.sh script to install them automatically. I've updated the conversion guide to include this step.