Image signature regression now uses losslessly encoded images.
When tinkering with crate versions, I got failures in the image_signature_regression test. I tracked this down to a version change in a transitive dependency, zune-jpeg, which was giving slightly different output when decoding the test images. Going from zune-jpeg 0.4.14 to 0.4.15 causes the test to fail.

Turns out this isn't necessarily wrong, as different JPEG decoders can give slightly different results due to factors like floating-point round-off error. However, it does make the regression test less useful, as a failure could either mean an bug in signature generation or a change in decoder implementation.

My solution is to use losslessly encoded images for the regression, as a correct decoder should always produce byte-for-byte identical images.