Image signature regression test now uses grayscale images.
This test started failing when updating depedencies, which turned out to be because the grayscale conversion implementation changed slightly. This isn't exactly an error on the part of the library, because even simple reordering of arithmetic operations can change floating-point round-off error. Therefore, the best solution is to only use images with are stored using an 8-bit grayscale format so that the conversion to grayscale is error-free.