Post data files are now split into subdirectories.
This was done because very large directories can become unwieldy to work with and eventually you run into filesystem limits on the number of files in a single directory.

Since a server migration is needed for this, I'm also using this opportunity to fix other issues that require migration like the bugged content filename hashing, the incorrect computing of Blake3 checksums, and a redundant index that needed removal.